added auth toke to picasso
